+++ config/smsConfigsLoader.js
@@ -11,8 +11,8 @@ app.ConfigName = {
 
 var conn = require('./connectionConfig');
 var configModelArray = [
-  rootRequire('api/models/donorsChooseBotModel')(conn),
-  rootRequire('api/models/donorsChooseMocoCampaignModel')(conn),
+  rootRequire('api/models/DonorsChooseBot')(conn),
+  rootRequire('api/models/DonorsChooseMoCoCampaign')(conn),
   rootRequire('api/legacy/ds-routing/config/startCampaignTransitionsConfigModel')(conn),
   rootRequire('api/legacy/ds-routing/config/yesNoPathsConfigModel')(conn),
   rootRequire('api/legacy/reportback/reportbackConfigModel')(conn),
